interface-based design

+++ src/main/java/definitions/structures/abstr/algebra/groups/DiscreetGroup.java
@@ -26,5 +26,5 @@ public interface DiscreetGroup extends DiscreetMonoid, Group {
 	 * @param representant the representant of the element.
 	 * @return the corresponding monoid element.
 	 */
-	Element get(Integer representant);
+	Element get(Double representant);
 }
