+++ src/main/java/definitions/structures/abstr/algebra/semigroups/DiscreetSemiGroup.java
@@ -9,9 +9,9 @@ public interface DiscreetSemiGroup extends SemiGroup {
 	 * @param representant the representant of the element.
 	 * @return the corresponding monoid element.
 	 */
-	Element get(Integer representant);
+	Element get(Double representant);
 
-	default Integer getRepresentant(Element first) {
+	default Double getRepresentant(Element first) {
 		return first.getRepresentant();
 	};
 }
