+++ src/main/java/definitions/structures/abstr/algebra/semigroups/Element.java
@@ -2,7 +2,8 @@ package definitions.structures.abstr.algebra.semigroups;
 
 public interface Element{
 
-	default Integer getRepresentant() {
+	default Double getRepresentant() {
 		return null;
 	}
+
 }
